refactor
remove dependency on Long, we have extracted the methods we need (and
fixed a bug in shr).
This makes the source MUCH more readable.
Clean up the API for bit board
add orB and andB (bit board or/and) which accepts a bit board argument,
and or/and which accept a 32-bit unsigned int and expands it to a
64-bit via duplication.